backend/requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
pydantic==2.9.2
pydantic-settings==2.6.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.17
python-dotenv==1.0.1
requests==2.32.3
sentence-transformers==3.3.1
openai==1.55.3
PyPDF2==3.0.1
python-docx==1.1.2
pandas==2.2.3
numpy~=1.26.0
scikit-learn==1.5.2
langchain==0.1.20
langchain-openai==0.1.7
supabase==2.3.4
fpdf2==2.7.9
reportlab==4.1.0
